ignore declaration when counting variable usage

+++ src/test/java/com/github/mauricioaniche/ck/metric/ParameterCountTest.java
@@ -20,33 +20,33 @@ public class ParameterCountTest  extends BaseTest{
 	public void justDeclaration() {
 		CKNumber a = report.get("pcount.A");
 		Map<String, Integer> variablesUsageA = a.getMethod("a/0").get().getVariablesUsage();
-		Assert.assertEquals(new Integer(1), variablesUsageA.get("a"));
-		Assert.assertEquals(new Integer(1), variablesUsageA.get("b"));
+		Assert.assertEquals(new Integer(0), variablesUsageA.get("a"));
+		Assert.assertEquals(new Integer(0), variablesUsageA.get("b"));
 	}
 
 	@Test
 	public void normalUse() {
 		CKNumber a = report.get("pcount.A");
 		Map<String, Integer> variablesUsageA = a.getMethod("b/0").get().getVariablesUsage();
-		Assert.assertEquals(new Integer(5), variablesUsageA.get("a")); // a appears 5 times
-		Assert.assertEquals(new Integer(3), variablesUsageA.get("b"));
+		Assert.assertEquals(new Integer(4), variablesUsageA.get("a")); // a appears 5 times
+		Assert.assertEquals(new Integer(2), variablesUsageA.get("b"));
 	}
 
 	@Test
 	public void passingVariableToMethod() {
 		CKNumber a = report.get("pcount.A");
 		Map<String, Integer> variablesUsageA = a.getMethod("bb/0").get().getVariablesUsage();
-		Assert.assertEquals(new Integer(2), variablesUsageA.get("a"));
+		Assert.assertEquals(new Integer(1), variablesUsageA.get("a"));
 	}
 
 	@Test
 	public void methodParameters() {
 		CKNumber a = report.get("pcount.A");
 		Map<String, Integer> variablesUsageA = a.getMethod("c/1[int]").get().getVariablesUsage();
-		Assert.assertEquals(new Integer(1), variablesUsageA.get("a"));
+		Assert.assertEquals(new Integer(0), variablesUsageA.get("a"));
 
 		Map<String, Integer> variablesUsageB = a.getMethod("d/2[int,int]").get().getVariablesUsage();
-		Assert.assertEquals(new Integer(5), variablesUsageB.get("a"));
-		Assert.assertEquals(new Integer(1), variablesUsageB.get("b"));
+		Assert.assertEquals(new Integer(4), variablesUsageB.get("a"));
+		Assert.assertEquals(new Integer(0), variablesUsageB.get("b"));
 	}
 }
